Restore the initial scope stack when catching an exception
This resets the scope stack to func_scope (the initial scope stack at
the start of the function), essentially removing all scopes that have
been added during the execution of the function. This fixes the
ability to resolve static properties after catching an exception
(static properties live in class objects which are part of
func_scope).